♻️ refacto: add builder/runner abstractions
