+++ pkg/builder/openapi.go
@@ -1,4 +1,4 @@
-package openapi
+package builder
 
 import (
 	"bufio"
